# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Common/Tasks"
                ],
                "summary": "Get a list of tasks",
                "description": "Retrieve a list of tasks, optionally filtered by status and paginated.",
                "operationId": "39f21f05e9be3bfad3d0a76c5744cab5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Comma-separated list of status IDs to filter tasks by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of tasks per page for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Common/Tasks"
                ],
                "summary": "Get a specific task by ID",
                "description": "Retrieve a specific task by its ID.",
                "operationId": "683b0b8d11af3e76ad1b0733c71fc823",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "put": {
                "tags": [
                    "Common/Tasks"
                ],
                "summary": "Update a specific task",
                "description": "Update the status of a specific task.",
                "operationId": "d4f6173d6c11f1f57fb3d2a4b2d5f790",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id"
                                ],
                                "properties": {
                                    "status_id": {
                                        "description": "ID of the new status",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/task/tasks/{task}": {
            "get": {
                "tags": [
                    "Task/Task"
                ],
                "summary": "Get a specific task",
                "operationId": "ed1a158b66e4869a1b414eaa5115a059",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error fetching task"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TaskResource": {
                "title": "Task Resource",
                "description": "Resource schema for a task",
                "properties": {
                    "id": {
                        "description": "Unique identifier of the task",
                        "type": "integer",
                        "example": 1
                    },
                    "work_order": {
                        "description": "Work order number",
                        "type": "string",
                        "example": "WO123456"
                    },
                    "customer_name": {
                        "description": "Name of the customer",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "street": {
                        "description": "Street address",
                        "type": "string",
                        "example": "123 Elm Street"
                    },
                    "city": {
                        "description": "City",
                        "type": "string",
                        "example": "Springfield"
                    },
                    "zip": {
                        "description": "Zip code",
                        "type": "string",
                        "example": "12345"
                    },
                    "authorization": {
                        "description": "Authorization details",
                        "type": "string",
                        "example": "Authorization for work"
                    },
                    "insurance": {
                        "$ref": "#/components/schemas/InsuranceResourceCommon"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TaskStatusResourceCommon"
                    },
                    "created_at": {
                        "description": "Creation timestamp of the task",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-28T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the task",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-28T12:34:56Z"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NoteResource"
                        }
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CustomerPhoneResource"
                        }
                    },
                    "files": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileResource"
                        }
                    },
                    "appliance_lists": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ApplianceListResourceCommon"
                        }
                    },
                    "received_payments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReceivedPaymentResource"
                        }
                    },
                    "technicians": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TechnicianResourceDispatch"
                        }
                    },
                    "parts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PartResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "ApplianceList": {
                "description": "Model representing an appliance list entry",
                "required": [
                    "serial_number",
                    "model_number",
                    "brand",
                    "appliance_id",
                    "task_id"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the appliance list entry",
                        "type": "integer"
                    },
                    "serial_number": {
                        "description": "Serial number of the appliance",
                        "type": "string"
                    },
                    "model_number": {
                        "description": "Model number of the appliance",
                        "type": "string"
                    },
                    "brand": {
                        "description": "Brand of the appliance",
                        "type": "string"
                    },
                    "dop": {
                        "description": "Date of purchase",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "appl_note": {
                        "description": "Additional note about the appliance",
                        "type": "string",
                        "nullable": true
                    },
                    "appliance_id": {
                        "description": "ID of the appliance type",
                        "type": "integer"
                    },
                    "task_id": {
                        "description": "ID of the task associated with this appliance",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Timestamp when the entry was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the entry was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "task": {
                        "$ref": "#/components/schemas/Task"
                    },
                    "appliance": {
                        "$ref": "#/components/schemas/Appliance"
                    }
                },
                "type": "object"
            },
...
            "CustomerPhone": {
                "description": "Model representing a customer's phone number",
                "required": [
                    "task_id",
                    "phone_number"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the customer phone entry",
                        "type": "integer"
                    },
                    "task_id": {
                        "description": "ID of the task associated with this phone number",
                        "type": "integer"
                    },
                    "phone_number": {
                        "description": "Customer's phone number",
                        "type": "string"
                    },
                    "note": {
                        "description": "Additional note regarding the phone number",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp when the entry was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the entry was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "task": {
                        "$ref": "#/components/schemas/Task"
                    }
                },
                "type": "object"
            },
...
            "Insurance": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "Insurance's ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Insurance's name",
                        "type": "string",
                        "example": "ACME Supplies"
                    },
                    "created_at": {
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time of the last update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Note": {
                "description": "Model representing a note attached to a task",
                "required": [
                    "text",
                    "user_id",
                    "task_id"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the note",
                        "type": "integer"
                    },
                    "text": {
                        "description": "Text content of the note",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID of the user who created the note",
                        "type": "integer"
                    },
                    "task_id": {
                        "description": "ID of the task to which the note is attached",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Timestamp when the note was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the note was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "task": {
                        "$ref": "#/components/schemas/Task"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "title": "Task",
                "description": "Task model",
                "required": [
                    "work_order",
                    "customer_name",
                    "status_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the task",
                        "type": "integer"
                    },
                    "work_order": {
                        "description": "Work order number",
                        "type": "string"
                    },
                    "customer_name": {
                        "description": "Name of the customer",
                        "type": "string"
                    },
                    "street": {
                        "description": "Street address of the customer",
                        "type": "string"
                    },
                    "city": {
                        "description": "City of the customer",
                        "type": "string"
                    },
                    "zip": {
                        "description": "ZIP code of the customer",
                        "type": "string"
                    },
                    "authorization": {
                        "description": "Authorization code",
                        "type": "string"
                    },
                    "insurance_id": {
                        "description": "ID of the related insurance",
                        "type": "integer"
                    },
                    "status_id": {
                        "description": "ID of the task status",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Timestamp of when the task was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of when the task was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TaskStatus"
                    },
                    "notes": {
                        "description": "List of notes associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Note"
                        }
                    },
                    "insurance": {
                        "$ref": "#/components/schemas/Insurance"
                    },
                    "phones": {
                        "description": "List of customer phones associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CustomerPhone"
                        }
                    },
                    "applianceLists": {
                        "description": "List of appliance lists associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ApplianceList"
                        }
                    },
                    "files": {
                        "description": "List of files associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskFile"
                        }
                    },
                    "taskBilling": {
                        "$ref": "#/components/schemas/TaskBilling"
                    },
                    "technicians": {
                        "description": "List of technicians associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "parts": {
                        "description": "List of parts associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Part"
                        }
                    },
                    "receivedPayments": {
                        "description": "List of received payments associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReceivedPayment"
                        }
                    },
                    "statusChanges": {
                        "description": "List of status changes associated with the task",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StatusChange"
                        }
                    }
                },
                "type": "object"
            },
...
            "TaskFile": {
                "required": [
                    "task_id",
                    "file_path",
                    "file_name"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the task file",
                        "type": "integer"
                    },
                    "task_id": {
                        "description": "ID of the task associated with the file",
                        "type": "integer"
                    },
                    "file_path": {
                        "description": "Path to the file in storage",
                        "type": "string"
                    },
                    "file_name": {
                        "description": "Original name of the file",
                        "type": "string"
                    },
                    "file_note": {
                        "description": "Optional note related to the file",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the file record was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the file record was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "task": {
                        "$ref": "#/components/schemas/Task"
                    }
                },
                "type": "object"
            },
            "TaskStatus": {
                "description": "Model representing status of a task",
                "required": [
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the status",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the status",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the status was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the status was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "tasks": {
                        "$ref": "#/components/schemas/Task"
                    },
                    "statusChanges": {
                        "$ref": "#/components/schemas/StatusChange"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}